% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanfields.R
\name{cleanfields}
\alias{cleanfields}
\title{cleanfields}
\usage{
cleanfields(sflayer)
}
\arguments{
\item{sflayer}{sf object}
}
\value{
it returns the same layer with the corrections over the special characters
}
\description{
This cleans some text fields where special characters like ends of line.
This is accesory function for PBADownloader
}
\examples{
\dontrun{
x <- st_read(sflayer)
cleanlayer <- cleanfields(x)
}
}
