% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedownload.R
\name{filedownload}
\alias{filedownload}
\title{filedownload}
\usage{
filedownload(url, filename)
}
\arguments{
\item{url}{url of the file (currently a .zip)}

\item{filename}{name of the downloaded file}
}
\value{
it returns the error or warning in case it has been producen.
Otherwise it just download the file into the desired filename directory
}
\description{
This is an accesory function that catch error when downloading zip files
It is OS dependent, using "curl" in Linux and "auto" in Windows.
}
\examples{
\dontrun{
# Download zip with municipality concello (in case it does not exist)
tempzip <- paste0(temp, concello, '.zip')
if(!file.exists(tempzip)){
                getzip <- filedownload(URLencode(ziplink), tempzip)
                } else {
                        cat("Not downloading PBA --> ", concello, "--> already exists")
                        }
}
}
