% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderpath.R
\name{folderpath}
\alias{folderpath}
\title{folderpath}
\usage{
folderpath(path)
}
\arguments{
\item{path}{text dir}
}
\value{
path of a folder or file
}
\description{
Accesory function to the the path of a folder or file
}
\examples{
\dontrun{
folderpath('./dir1/dir2/')
> './dir1/'
}
}
