% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parPBADownloader.R
\name{parPBADownloader}
\alias{parPBADownloader}
\title{parPBADownloader}
\usage{
parPBADownloader(concellos, outdir, ncores = 5)
}
\arguments{
\item{concellos}{Municipality name based on the pbaurls (availables in data)}

\item{outdir}{directory where file structure of downloads and products will be created}

\item{ncores}{number of cores used for the process}
}
\value{
it returns nothing but the download and folder structure.
}
\description{
función to do the same as PBADownloader in parallel
(careful with restrictions...max 5-6 simultaneous downloads)
OS independent function (written just for windows and linux)
}
\examples{
\dontrun{
# PARAMETERS
data("pbaurls")
outdir <- '../02_OUTPUT/'

pbaurls$Concello # show municipalities available
concellos <- c("Paderne", "Pol","A Peroxa", "Sarria")
parPBADownloader(concellos, outdir, ncores = 5)

# DOWNLOAD ALL IN PARALLEL
concellos <- pbaurls$Concello
parPBADownloader(concellos, outdir, ncores = 6)
}
}
