% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBADownloader.R
\name{PBADownloader}
\alias{PBADownloader}
\title{PBADownloader}
\usage{
PBADownloader(concello, outdir)
}
\arguments{
\item{concello}{Municipality name based on the pbaurls (availables in data)}

\item{outdir}{directory where file structure of downloads and products will be created}
}
\value{
it returns nothing but the download and folder structure.
}
\description{
this is the main function of the package. It downloads the .zip file for a Municipality (concello)
and extracts from it the vectorial information to stored into a single .gpkg file.
This are the particularities:
Temporal data of extractions will be remove.
Original .zip file will be kept in folder. This serves as security copy and avoid recurrent downloads.
}
\examples{
\dontrun{
# PARAMETERS
# read data
data("pbaurls")
outdir <- '../02_OUTPUT/'

# Single execution
PBADownloader("Beade", outdir)

# Single execution in a serie of municipalities (on by one)
concellos <- c("Paderne", "Taboada", "Parada De Sil", "Vedra")

lapply(concellos, PBADownloader, outdir)

# Download all
concellos <- pbaurls$Concello
lapply(concellos, PBADownloader, outdir)
}
}
