% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBAmixer.R
\name{PBAmixer}
\alias{PBAmixer}
\title{PBAmixer}
\usage{
PBAmixer(outdir)
}
\arguments{
\item{outdir}{directory where PBADownloader have created the structure (it should be the same outdir for both functions)}
}
\value{
Funcition to read the .gpkg files created and merge layers into a superior file
}
\description{
Funcition to read the .gpkg files created and merge layers into a superior file
}
\examples{
\dontrun{
# PARAMETERS
data("pbaurls")
outdir <- '../02_OUTPUT/'
PBAmixer(outdir)
}
}
