% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbaurls.R
\name{pbaurls}
\alias{pbaurls}
\title{pbaurls}
\usage{
pbaurls(
  layer = "./data/base/Concellos_IGN/Concellos_IGN.shp",
  attrname = "NomeMAY",
 
    mainurl = "visorgis.cmati.xunta.es/cdix/descargas/PlanBasicoAutonomico/2020/concellos_ZIPS/",
  PBAprefix = "PBA_2020_"
)
}
\arguments{
\item{layer}{Concellos_IGN.shp from IET that serves as base for Concellos Names}

\item{attrname}{atrribute name used in the function to capture Concellos Names}

\item{mainurl}{base url where zips are stored}

\item{PBAprefix}{text piece that goes before Concello to construct the zip url}
}
\value{
data.frame with 2 fields Concello|URL . This gives the user the
list of all Concellos available (and the right way to write their names...)
}
\description{
function to create the fixed pbaurls (users don't have to use
 this unless they are aware of a change in the url composition of PBA or Concello change of name)
}
\examples{
\dontrun{
# create pbaurls (a priori it shouldnt change...)
pbaurls <- pbaurls() # create table
save(pbaurls, file = "./data/pbaurls.RData") # save table
}
}
